refactor(SocketManager): move socket manager into managers
