+++ src/system/managers/socket/index.js
@@ -1,0 +1,13 @@
+'use strict'
+
+/**
+ * Module Dependencies
+ * @ignore
+ */
+const SocketManager = require('./SocketManager')
+
+/**
+ * Export
+ * @ignore
+ */
+module.exports = SocketManager
